# backend/src/main/openapi/faktura/api_v1.yaml
openapi: 3.0.2
info:
  title: faktura
  version: 1.0.0
...
paths:
...
  /customer:
    get:
      tags:
        - customer
      operationId: getCustomers
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'
    post:
      tags:
        - customer
      operationId: saveCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customer/{id}:
    get:
      tags:
        - customer
      summary: Find by ID.
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    delete:
      tags:
        - customer
      summary: Deletes a entity.
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
  /earning:
    get:
      tags:
        - earning
      operationId: getEarnings
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EarningPage'
    post:
      tags:
        - earning
      operationId: saveEarning
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Earning'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Earning'
  /earning/{id}:
    get:
      tags:
        - earning
      summary: Find by ID.
      operationId: getEarningById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Earning'
    delete:
      tags:
        - earning
      summary: Deletes a entity.
      operationId: deleteEarning
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
...
  /invoice:
    get:
      tags:
        - invoice
      operationId: getInvoices
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoicePage'
    post:
      tags:
        - invoice
      operationId: saveInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /invoice/{id}:
    get:
      tags:
        - invoice
      summary: Find by ID.
      operationId: getInvoiceById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
    delete:
      tags:
        - invoice
      summary: Deletes a entity.
      operationId: deleteInvoice
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
...
  /saleService:
    get:
      tags:
        - saleService
      operationId: getSaleServices
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleServicePage'
    post:
      tags:
        - saleService
      operationId: saveSaleService
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleService'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleService'
  /saleService/{id}:
    get:
      tags:
        - saleService
      summary: Find by ID.
      operationId: getSaleServiceById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleService'
    delete:
      tags:
        - saleService
      summary: Deletes a entity.
      operationId: deleteSaleService
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
  /setting:
    get:
      tags:
        - setting
      operationId: getSettings
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingPage'
    post:
      tags:
        - setting
      operationId: saveSetting
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Setting'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
  /setting/{id}:
    get:
      tags:
        - setting
      summary: Find by ID.
      operationId: getSettingById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
    delete:
      tags:
        - setting
      summary: Deletes a entity.
      operationId: deleteSetting
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
  /supplier:
    get:
      tags:
        - supplier
      operationId: getSuppliers
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierPage'
    post:
      tags:
        - supplier
      operationId: saveSupplier
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Supplier'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
  /supplier/{id}:
    get:
      tags:
        - supplier
      summary: Find by ID.
      operationId: getSupplierById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
    delete:
      tags:
        - supplier
      summary: Deletes a entity.
      operationId: deleteSupplier
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
  /timeEntry:
    get:
      tags:
        - timeEntry
      operationId: getTimeEntries
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntryPage'
    post:
      tags:
        - timeEntry
      operationId: saveTimeEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeEntry'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'
  /timeEntry/{id}:
    get:
      tags:
        - timeEntry
      summary: Find by ID.
      operationId: getTimeEntryById
      parameters:
        - name: id
          in: path
          description: ID of entity
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'
    delete:
      tags:
        - timeEntry
      summary: Deletes a entity.
      operationId: deleteTimeEntry
      parameters:
        - name: id
          in: path
          description: Id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
components:
  schemas:
    Pageable:
      description: Pageable query parameters
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
    Page:
      description: Page response
      type: object
      properties:
        number:
          type: integer
        size:
          type: integer
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the account
        description:
          type: string
          maxLength: 512
          description: Description of the account
        noCharging:
          type: boolean
          description: Flag indicating if charging is disabled for this account
      required:
        - id
        - description
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the payment method
        description:
          type: string
          maxLength: 512
          description: Description of the payment method
      required:
        - id
        - description
...
    SupplierPage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/SupplierList'
    SupplierList:
      description: List of Suppliers
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Supplier'
    Supplier:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the supplier
        searchName:
          type: string
          maxLength: 512
          description: Search name for the supplier
        name:
          type: string
          maxLength: 512
          description: Full name of the supplier
      required:
        - id
        - searchName
        - name
    SettingPage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/SettingList'
    SettingList:
      description: List of Settings
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Setting'
    Setting:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the setting
        key:
          type: string
          maxLength: 64
          description: Key identifier for the setting
        value:
          type: string
          maxLength: 1024
          description: Value of the setting
      required:
        - id
        - key
        - value
...
    InvoicePage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/InvoiceList'
    InvoiceList:
      description: List of Invoices
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
    Invoice:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the invoice
        invoiceNumber:
          type: integer
          format: int64
          description: Invoice number
        subject:
          type: string
          maxLength: 1024
          description: Subject of the invoice
        createdOn:
          type: string
          format: date
          description: Date when the invoice was created
        paidOn:
          type: string
          format: date
          description: Date when the invoice was paid
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        customer:
          $ref: '#/components/schemas/Customer'
        servicePeriod:
          type: string
          maxLength: 1024
          description: Period of service covered by the invoice
        revision:
          type: string
          maxLength: 1024
          description: Revision information for the invoice
      required:
        - id
        - subject
        - createdOn
        - customerId
    CustomerPage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/CustomerList'
    CustomerList:
      description: List of Customers
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the customer
        searchName:
          type: string
          maxLength: 256
          description: Search name for the customer
        name:
          type: string
          maxLength: 512
          description: Full name of the customer
        addressLine1:
          type: string
          maxLength: 512
          description: First line of customer address
        addressLine2:
          type: string
          maxLength: 512
          description: Second line of customer address
        city:
          type: string
          maxLength: 128
          description: City of the customer
        postalCode:
          type: string
          maxLength: 32
          description: Postal code of the customer
        taxIdentificationNumber:
          type: string
          maxLength: 64
          description: Sales tax identifier for the customer
        email:
          type: string
          maxLength: 256
          description: Email address of the customer
        phone:
          type: string
          maxLength: 256
          description: Phone number of the customer
        hourlyRate:
          type: number
          format: double
          description: Hourly rate charged to the customer
        countryCode:
          type: string
          maxLength: 3
          description: Country code of the customer
        paymentInfo:
          type: string
          maxLength: 1024
          description: Payment information for the customer
      required:
        - id
        - searchName
        - name
        - addressLine1
        - city
        - postalCode
        - countryCode
    EarningPage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/EarningList'
    EarningList:
      description: List of Earnings
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Earning'
    Earning:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the earning
        description:
          type: string
          maxLength: 1024
          description: Description of the earning
        account:
          $ref: '#/components/schemas/Account'
        priceNet:
          type: number
          format: double
          description: Net price of the earning
        paidOn:
          type: string
          format: date
          description: Date when the earning was paid
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        supplier:
          $ref: '#/components/schemas/Supplier'
        note:
          type: string
          maxLength: 1024
          description: Additional notes about the earning
      required:
        - id
        - description
...
    SaleServicePage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/SaleServiceList'
    SaleServiceList:
      description: List of SalesServices
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SaleService'
    SaleService:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the sales service
        customer:
          $ref: '#/components/schemas/Customer'
        hourlyRate:
          type: number
          format: double
          description: Hourly rate for the service
        suppliedOn:
          type: string
          format: date
          description: Date when the service was supplied
        hours:
          type: number
          format: float
          description: Number of hours of service provided
        description:
          type: string
          maxLength: 1024
          description: Description of the sales service
        invoice:
          $ref: '#/components/schemas/Invoice'
        timeInfo:
          type: string
          maxLength: 256
          description: Time information for the service
        salesNet:
          type: number
          format: double
          description: Net sales amount
      required:
        - id
        - customerId
        - hourlyRate
        - suppliedOn
        - hours
        - description
    TimeEntryPage:
      description: typed Page
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - $ref: '#/components/schemas/TimeEntryList'
    TimeEntryList:
      description: List of TimeEntries
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TimeEntry'
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Primary key identifier for the time entry
        customer:
          $ref: '#/components/schemas/Customer'
        startedOn:
          type: string
          format: date-time
          description: Date and time when the entry started
        endedOn:
          type: string
          format: date-time
          description: Date and time when the entry ended
        description:
          type: string
          maxLength: 1024
          description: Description of the time entry
        saleService:
          $ref: '#/components/schemas/SaleService'
      required:
        - id
        - customerId
        - startedOn
        - description